Separating the platform metadata retrieval function from within the logger module.
