+++ studio/app/common/core/logger.py
@@ -1,16 +1,15 @@
 import hashlib
-import json
 import logging
 import logging.config
 import os
 import platform
 import traceback
-import urllib.request
 from contextvars import ContextVar
 from typing import Optional
 
 import yaml
 
+from studio.app.common.core.platform_metadata import ECS_TASK_ID
 from studio.app.common.core.mode import MODE
 from studio.app.common.core.utils.datetime_utils import get_current_datetime_formatted
 from studio.app.dir_path import DIRPATH
@@ -22,48 +21,7 @@ _client_id_context: ContextVar[Optional[str]] = ContextVar(
     LOGGING_CLIENT_ID_KEY, default=None
 )
 
-NO_ECS_TASK_DEFAULT = "local"
-NO_ECS_SERVICE_DEFAULT = "none"
-_ECS_METADATA_TIMEOUT = 2
 VALID_LOG_LEVELS = {"DEBUG", "INFO", "WARNING", "ERROR", "CRITICAL"}
-
-
-def _get_ecs_task_id() -> str:
-    """Fetch the short ECS task ID from the container metadata
-    endpoint (v4). Returns a default value outside ECS."""
-    meta_uri = os.environ.get("ECS_CONTAINER_METADATA_URI_V4")
-    if not meta_uri:
-        return NO_ECS_TASK_DEFAULT
-    try:
-        req = urllib.request.Request(f"{meta_uri}/task")
-        with urllib.request.urlopen(req, timeout=_ECS_METADATA_TIMEOUT) as resp:
-            data = json.loads(resp.read())
-        # TaskARN: arn:aws:ecs:region:account:task/cluster/id
-        task_arn = data.get("TaskARN", "")
-        return task_arn.rsplit("/", 1)[-1] if "/" in task_arn else NO_ECS_TASK_DEFAULT
-    except Exception:
-        return NO_ECS_TASK_DEFAULT
-
-
-def _get_ecs_service_name() -> str:
-    """Fetch the ECS service name from the container metadata
-    endpoint (v4). Returns a default value if not available."""
-    meta_uri = os.environ.get("ECS_CONTAINER_METADATA_URI_V4")
-    if not meta_uri:
-        return NO_ECS_SERVICE_DEFAULT
-    try:
-        req = urllib.request.Request(f"{meta_uri}/task")
-        with urllib.request.urlopen(req, timeout=_ECS_METADATA_TIMEOUT) as resp:
-            data = json.loads(resp.read())
-        # ServiceName is only available for tasks started by a service
-        service_name = data.get("ServiceName", "")
-        return service_name if service_name else NO_ECS_SERVICE_DEFAULT
-    except Exception:
-        return NO_ECS_SERVICE_DEFAULT
-
-
-ECS_TASK_ID: str = _get_ecs_task_id()
-ECS_SERVICE_NAME: str = _get_ecs_service_name()
 
 
 class LoggingConfigHelper:
